+++ NEWS.md
@@ -17,7 +17,7 @@
 * feta: Added helper `cat_cli()`.
 * fix: Default `dicts_suggest = NULL` for `dictionary_get_item()` and `dictionary_retrieve_item()` for backward compatibility.
 * fix: Wrong assert in `dictionary_sugar_inc_get`.
-* feat: Functions `warningf()` and `stopf()` now have a `class` argument and also add the additional class `mlr3warning` and `mlr3error`, respectively.
+* feat: Functions `warningf()` and `stopf()` now have a `class` argument and also add the additional class `Mlr3Warning` and `Mlr3Error`, respectively.
   The condition object now also includes the call.
 
 # mlr3misc 0.16.0